Add StreamingServer
New ScriptStreamingServer is a TCP server which accepts
connection and write the plain text from game window into the socket.
It is a read-only connection, where clients want to get the stream
of the game text.
Introduced setting for enabling and setting the port of this server
in the Script Settings dialog.